feat: Refactor project display and types
Updates the project data structure to use icons instead of images and introduces a new `OtherProject` type. Also refactors the `Projects` component to handle featured and other projects, with the ability to toggle visibility. The `ProjectCard` component is updated to accommodate the new design and data structure. Preload images are removed from `index.html`.